+++ src/day_05.rs
@@ -16,7 +16,7 @@ fn convert_seeds(seeds: Vec<u64>, converters: Vec<[u64; 3]>) -> Vec<u64> {
         .collect::<Vec<u64>>();
 }
 
-pub fn solve(input: String) -> (u32, u32) {
+pub fn solve(input: String) -> (u64, u64) {
     let lines = input.lines().collect::<Vec<&str>>();
 
     let input_seeds = lines[0]
