+++ src/day_06.rs
@@ -15,7 +15,7 @@ fn solve_race(input: (usize, usize)) -> usize {
     return res;
 }
 
-pub fn solve(input: String) -> (u32, u32) {
+pub fn solve(input: String) -> (u64, u64) {
     let lines = input.lines().collect::<Vec<&str>>();
 
     let first_line = lines
